chore: move polyfill import to shared directory
