+++ app/common/actions/Session.js
@@ -17,7 +17,7 @@ import {Web2Driver} from 'web2driver';
 import i18n from '../configs/i18next.renderer';
 import {APP_MODE} from '../constants/session-inspector';
 import CloudProviders from '../components/Session/CloudProviders.jsx';
-import {fs, ipcRenderer, util} from '../utils/polyfills';
+import {fs, ipcRenderer, util} from '../shared/polyfills';
 import {
   SAVED_SESSIONS,
   SERVER_ARGS,
@@ -78,7 +78,6 @@ export const SET_ADD_VENDOR_PREFIXES = 'SET_ADD_VENDOR_PREFIXES';
 export const SET_CAPABILITY_NAME_ERROR = 'SET_CAPABILITY_NAME_ERROR';
 export const SET_STATE_FROM_URL = 'SET_STATE_FROM_URL';
 export const SET_STATE_FROM_SAVED = 'SET_STATE_FROM_SAVED';
-
 
 const APPIUM_SESSION_FILE_VERSION = '1.0';
