+++ app/common/shared/polyfills.js
